# Improve CENTER split raycast performance