--- a/src/BVHConstructionContext.js
+++ b/src/BVHConstructionContext.js
@@ -207,10 +207,43 @@ export default class BVHConstructionContext {
 		// Center
 		if ( strategy === CENTER ) {
 
-			axis = getLongestEdgeIndex( bounds );
+			const triBounds = this.bounds;
+			const minSpread = new Float32Array( 3 ).fill( Infinity );
+			const maxSpread = new Float32Array( 3 ).fill( - Infinity );
+
+			for ( let i = offset * 6, l = ( offset + count ) * 6; i < l; i += 6 ) {
+
+				const cx = triBounds[ i ];
+				const cy = triBounds[ i + 2 ];
+				const cz = triBounds[ i + 4 ];
+
+				if ( cx < minSpread[ 0 ] ) minSpread[ 0 ] = cx;
+				if ( cx > maxSpread[ 0 ] ) maxSpread[ 0 ] = cx;
+
+				if ( cy < minSpread[ 1 ] ) minSpread[ 1 ] = cy;
+				if ( cy > maxSpread[ 1 ] ) maxSpread[ 1 ] = cy;
+
+				if ( cz < minSpread[ 2 ] ) minSpread[ 2 ] = cz;
+				if ( cz > maxSpread[ 2 ] ) maxSpread[ 2 ] = cz;
+
+			}
+
+			let curr = - Infinity;
+			for ( let i = 0; i < 3; i ++ ) {
+
+				let dist = maxSpread[ i ] - minSpread[ i ];
+				if ( dist > curr ) {
+
+					curr = dist;
+					axis = i;
+
+				}
+
+			}
+
 			if ( axis !== - 1 ) {
 
-				pos = ( bounds[ axis + 3 ] + bounds[ axis ] ) / 2;
+				pos = ( maxSpread[ axis ] + minSpread[ axis ] ) / 2;
 
 			}