--- a/src/BVHConstructionContext.js
+++ b/src/BVHConstructionContext.js
@@ -86,7 +86,9 @@ export default class BVHConstructionContext {
 
 	}
 
-	// computes the union of the bounds of all of the given triangles and puts the resulting box in target
+	// computes the union of the bounds of all of the given triangles and puts the resulting box in target. If
+	// centroidTarget is provided then a bounding box is computed for the centroids of the triangles, as well.
+	// These are computed together to avoid redundant accesses to bounds array.
 	getBounds( offset, count, target, centroidTarget = null ) {
 
 		let minx = Infinity;
@@ -103,8 +105,8 @@ export default class BVHConstructionContext {
 		let cmaxy = - Infinity;
 		let cmaxz = - Infinity;
 
+		const includeCentroid = centroidTarget !== null;
 		const bounds = this.bounds;
-		const includeCentroids = centroidTarget !== null;
 		for ( let i = offset * 6, end = ( offset + count ) * 6; i < end; i += 6 ) {
 
 			const cx = bounds[ i + 0 ];
@@ -113,8 +115,8 @@ export default class BVHConstructionContext {
 			const rx = cx + hx;
 			if ( lx < minx ) minx = lx;
 			if ( rx > maxx ) maxx = rx;
-			if ( includeCentroids && cx < cminx ) cminx = cx;
-			if ( includeCentroids && cx > cmaxx ) cmaxx = cx;
+			if ( includeCentroid && cx < cminx ) cminx = cx;
+			if ( includeCentroid && cx > cmaxx ) cmaxx = cx;
 
 			const cy = bounds[ i + 2 ];
 			const hy = bounds[ i + 3 ];
@@ -122,8 +124,8 @@ export default class BVHConstructionContext {
 			const ry = cy + hy;
 			if ( ly < miny ) miny = ly;
 			if ( ry > maxy ) maxy = ry;
-			if ( includeCentroids && cy < cminy ) cminy = cy;
-			if ( includeCentroids && cy > cmaxy ) cmaxy = cy;
+			if ( includeCentroid && cy < cminy ) cminy = cy;
+			if ( includeCentroid && cy > cmaxy ) cmaxy = cy;
 
 			const cz = bounds[ i + 4 ];
 			const hz = bounds[ i + 5 ];
@@ -131,8 +133,8 @@ export default class BVHConstructionContext {
 			const rz = cz + hz;
 			if ( lz < minz ) minz = lz;
 			if ( rz > maxz ) maxz = rz;
-			if ( includeCentroids && cz < cminz ) cminz = cz;
-			if ( includeCentroids && cz > cmaxz ) cmaxz = cz;
+			if ( includeCentroid && cz < cminz ) cminz = cz;
+			if ( includeCentroid && cz > cmaxz ) cmaxz = cz;
 
 		}
 
@@ -144,7 +146,7 @@ export default class BVHConstructionContext {
 		target[ 4 ] = maxy;
 		target[ 5 ] = maxz;
 
-		if ( includeCentroids ) {
+		if ( includeCentroid ) {
 
 			centroidTarget[ 0 ] = cminx;
 			centroidTarget[ 1 ] = cminy;
@@ -155,6 +157,43 @@ export default class BVHConstructionContext {
 			centroidTarget[ 5 ] = cmaxz;
 
 		}
+
+	}
+
+	// A stand alone function for retrieving the centroid bounds.
+	getCentroidBounds( offset, count, centroidTarget ) {
+
+		let cminx = Infinity;
+		let cminy = Infinity;
+		let cminz = Infinity;
+		let cmaxx = - Infinity;
+		let cmaxy = - Infinity;
+		let cmaxz = - Infinity;
+
+		const bounds = this.bounds;
+		for ( let i = offset * 6, end = ( offset + count ) * 6; i < end; i += 6 ) {
+
+			const cx = bounds[ i + 0 ];
+			if ( cx < cminx ) cminx = cx;
+			if ( cx > cmaxx ) cmaxx = cx;
+
+			const cy = bounds[ i + 2 ];
+			if ( cy < cminy ) cminy = cy;
+			if ( cy > cmaxy ) cmaxy = cy;
+
+			const cz = bounds[ i + 4 ];
+			if ( cz < cminz ) cminz = cz;
+			if ( cz > cmaxz ) cmaxz = cz;
+
+		}
+
+		centroidTarget[ 0 ] = cminx;
+		centroidTarget[ 1 ] = cminy;
+		centroidTarget[ 2 ] = cminz;
+
+		centroidTarget[ 3 ] = cmaxx;
+		centroidTarget[ 4 ] = cmaxy;
+		centroidTarget[ 5 ] = cmaxz;
 
 	}
 
@@ -231,7 +270,7 @@ export default class BVHConstructionContext {
 
 	}
 
-	getOptimalSplit( bounds, offset, count, strategy ) {
+	getOptimalSplit( bounds, centroidBounds, offset, count, strategy ) {
 
 		let axis = - 1;
 		let pos = 0;
@@ -239,30 +278,12 @@ export default class BVHConstructionContext {
 		// Center
 		if ( strategy === CENTER ) {
 
-			const triBounds = this.bounds;
-			let minx = Infinity;
-			let miny = Infinity;
-			let minz = Infinity;
-			let maxx = - Infinity;
-			let maxy = - Infinity;
-			let maxz = - Infinity;
-
-			for ( let i = offset * 6, l = ( offset + count ) * 6; i < l; i += 6 ) {
-
-				const cx = triBounds[ i ];
-				const cy = triBounds[ i + 2 ];
-				const cz = triBounds[ i + 4 ];
-
-				if ( cx < minx ) minx = cx;
-				if ( cx > maxx ) maxx = cx;
-
-				if ( cy < miny ) miny = cy;
-				if ( cy > maxy ) maxy = cy;
-
-				if ( cz < minz ) minz = cz;
-				if ( cz > maxz ) maxz = cz;
-
-			}
+			const minx = centroidBounds[ 0 ];
+			const miny = centroidBounds[ 1 ];
+			const minz = centroidBounds[ 2 ];
+			const maxx = centroidBounds[ 3 ];
+			const maxy = centroidBounds[ 4 ];
+			const maxz = centroidBounds[ 5 ];
 
 			const widthx = maxx - minx;
 			const widthy = maxy - miny;