Fix #121

Improves the CENTER split method by picking the center of the centroid bounds along the longest axis rather than the center of the node bounds to partition against. This fixes cases like in #121 where you might have a node that winds up containing tens of thousands of triangles because a naive CENTER split of the node doesn't effectively divide the triangles and quit traversing.

The triangle centroid bounds are computed in the same pass as the node bounding box to avoid redundant array accesses. This minimizes the overhead of generating the centroid bounds and along with the improvements from #126 there is basically no performance degradation in computing the BVH compared to the last v0.1.5 release.

#### Benchmarks

Here are the significant benchmark comparisions relative to master (including #126).

*Nominal Case*
```
memory                   : 9755.746 kb  -> 10517.942 kb
Compute BVH              : 93.272 ms    -> 103.724 ms
BVH Raycast              : 0.650 ms     -> 0.138 ms
First Hit Raycast        : 0.097 ms     -> 0.004 ms
IntersectsSphere         : 0.013 ms     -> 0.001 ms
```

And in the *Extreme Tower Geometry* case
```
memory                   : 548.162 kb -> memory: 9449.764 kb
tris min                 : 4          -> 3
tris max                 : 300649     -> 9
CENTER raycast           : 10.117 ms  -> 0.003 ms
```

---

@mqp I'm not sure if you're still interested in reviewing changes to this but I figured since you're using for a project it would be good to get your feedback. I think it results in a generally better tree structure that avoids corner cases with some geometry like mentioned above. There's also a pretty significant run time raycast performance improvement which is great but I know you mentioned in the past that you were more interested in improving the build time performance.

#42 may be the next thing I take a look at tackling which would enable the bounds to be generated a web worker.