@gkjohnson 2020-01-30
After running a few tests with this PR in my application I've actually noticed that while raycast performance is significantly improved and more consistent build performance has dropped a bit from ~30ms to ~80-90ms for my particular geometry. Granted this is geometry that had a very poorly constructed tree before with nodes that had tons of triangles in them so it's likely that this type of geometry had its build time affected the most. I'll have to perform a few more tests with some more well-constructed models that we've tested with in the past.