@gkjohnson 2020-01-30
After more testing with [this Pickle Rick model](https://sketchfab.com/3d-models/pickle-rick-d779bda72fe441c297836a197640a709) it looks like the problematic models with bad triangle splits take a bit longer to build but for more regular geometry there's a big improvement compared to `v0.1.5` so I'm thinking I'll merge this as it looks like a general improvement. Here are the build timing comparisons:

```
v0.1.5  : ~ 1100ms - 1250ms
master  : ~ 400ms  - 415ms
this PR : ~ 430ms  - 450ms
```

And the code used for the benchmark in latest chrome v79.0.3945.130.

```js
const loader = new GLTFLoader();
loader.load( '../pickle_rick/scene.gltf', res => {

    const geom = [];
    res.scene.traverse(c => c.isMesh && geom.push( c.geometry ) );

    runBenchmark(
        'BVH Build Time',
        () => {

            geom.forEach( g => {

                g.disposeBoundsTree();
                g.computeBoundsTree();

            } );

        },
        3000
    );

});
```